Move targetsToContinuousTarget to a new file in src/util.
