+++ packages/cursorless-engine/src/processTargets/TargetPipelineRunner.ts
@@ -19,6 +19,7 @@ import { ImplicitStage } from "./marks/ImplicitStage";
 import { ContainingTokenIfUntypedEmptyStage } from "./modifiers/ConditionalModifierStages";
 import { PlainTarget } from "./targets";
 import { uniqWithHash } from "../util/uniqWithHash";
+import {targetsToContinuousTarget} from "../util/targetsToContinuousTarget";
 
 export class TargetPipelineRunner {
   constructor(
@@ -277,7 +278,7 @@ function getExcludedScope(
   );
 }
 
-function calcIsReversed(anchor: Target, active: Target) {
+export function calcIsReversed(anchor: Target, active: Target) {
   if (anchor.contentRange.start.isAfter(active.contentRange.start)) {
     return true;
   }
@@ -295,32 +296,10 @@ function uniqTargets(array: Target[]): Target[] {
   );
 }
 
-function ensureSingleEditor(anchorTarget: Target, activeTarget: Target) {
+export function ensureSingleEditor(anchorTarget: Target, activeTarget: Target) {
   if (anchorTarget.editor !== activeTarget.editor) {
     throw new Error("Cannot form range between targets in different editors");
   }
-}
-
-export function targetsToContinuousTarget(
-  anchorTarget: Target,
-  activeTarget: Target,
-  excludeAnchor: boolean = false,
-  excludeActive: boolean = false,
-): Target {
-  ensureSingleEditor(anchorTarget, activeTarget);
-
-  const isReversed = calcIsReversed(anchorTarget, activeTarget);
-  const startTarget = isReversed ? activeTarget : anchorTarget;
-  const endTarget = isReversed ? anchorTarget : activeTarget;
-  const excludeStart = isReversed ? excludeActive : excludeAnchor;
-  const excludeEnd = isReversed ? excludeAnchor : excludeActive;
-
-  return startTarget.createContinuousRangeTarget(
-    isReversed,
-    endTarget,
-    !excludeStart,
-    !excludeEnd,
-  );
 }
 
 function targetsToVerticalTarget(
